requirements.txt
Pillow>=8.2.0
tqdm